requirements.txt
cfgv==3.3.1
contourpy==1.0.5
cycler==0.11.0
distlib==0.3.6
filelock==3.8.0
fonttools==4.37.4
graphviz==0.20.1
identify==2.5.6
kiwisolver==1.4.4
matplotlib==3.6.1
nodeenv==1.7.0
numpy==1.23.4
packaging==21.3
pandas==1.5.1
Pillow==9.2.0
platformdirs==2.5.2
pre-commit==2.20.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.5
PyYAML==6.0
six==1.16.0
toml==0.10.2
virtualenv==20.16.5
